Improve : Make DefaultAttackComponent and DefaultSkillAttackComponent Methods Virtual
in order to override you have to create ur own script and inherit the interface, and copy paste all the code that u dont wanna change to just change few lines..
Or u have to inherit the default component with "new" and also everything that calls that method has to be "new" otherwise it still uses the base class method.

This makes it much easier to change small stuff or all u need.