Modified UICharacterStats to be able to show stat using N0
Like the current logic implemented in the other UI scripts, the user is able to change the format of the float changing the string in editor to "N0, N2, etc". Its useful to have that for the stats too because some users would like to show for example HP 70 instead of HP 70.00.